Moved location of and changed CaseSensitiveTagHelperAttributeComparer to a more generic one
