+++ test/Microsoft.AspNetCore.Razor.Test/project.json
@@ -3,6 +3,7 @@
   "dependencies": {
     "dotnet-test-xunit": "2.2.0-*",
     "Microsoft.AspNetCore.Razor": "1.1.0-*",
+    "Microsoft.AspNetCore.Razor.Runtime": "1.1.0-*",
     "Microsoft.AspNetCore.Razor.Test.Sources": {
       "version": "1.1.0-*",
       "type": "build"
